Crash: clear compiler warnings
- Crash.ELFormat: cast both operands to UInt32 in the exception-ID hash (W1024 signed/unsigned).
- Crash.Reporter: use the non-deprecated TMultipartFormData.AddStream overload with explicit AOwnsStream=False (W1000).
- Crash.MacOS.MachExc: drop dead Result init in the never-returning ExcWatcherThread (H2077).
- Crash.LineNumbers: drop the redundant default Result on the Linux Load path; the except handler already sets Exception (H2077).
- Stealth.Base.CrashSetup: silence the constant-fold warning on the build-generated endpoint port comparison (W1022).

+++ Crash.ELFormat.pas
@@ -324,7 +324,7 @@ begin
   Msg := AReport.ExceptionMessage;
   Hash := UInt32(Loc.CodeAddress xor (Loc.CodeAddress shr 32));
   if Msg <> '' then
-    Hash := Hash xor UInt32(Length(Msg) * 2654435761);
+    Hash := Hash xor (UInt32(Length(Msg)) * UInt32(2654435761));
   Result := IntToHex(Hash, 8);
 end;
